feat(contracts): hardening cycles 2-4 — emergency withdraw, governance guards
Second and third-pass brutalist review findings:
- SC-5: governance resolution requires resolutionDeadline check
- SC-6: setMinParticipants only allows lowering (raising breaks active debates)
- FP5-F01: emergency withdraw from AWAITING_GOVERNANCE (prevents fund lock)
- FP-3: LMSR quantities store raw weighted amounts, not b-normalized
- SL-10: deriveDomain public→internal (save ABI exposure)
- SL-12: engagement root expiry on registration
- SL-14: MAX_DISTRICT_SLOTS uint256→uint8
- SM-7: re-registration dedup in AIEvaluationRegistry modelList
- Add minParticipants check in submitAIEvaluation
- Update debate market spec for TEE-attested evaluation model
- Add demo deploy script for Sepolia

+++ contracts/src/Constants.sol
@@ -4,4 +4,4 @@ pragma solidity 0.8.28;
 /// @notice Shared constants across the voter-protocol contract stack
 
 /// @dev Maximum district types a user can be registered in (20 defined + 4 overflow)
-uint256 constant MAX_DISTRICT_SLOTS = 24;
+uint8 constant MAX_DISTRICT_SLOTS = 24;
